Update workoutvidselector.py
Making the script take the To:/From: email addresses as system arguments that way the user doesn't have to edit the script to supply them. Also added a "workout" variable and updated the embedded HTML header (H1) tag to state what type of workouts are being focused on that day.